Add BSD compatibility queue.h
This creates a new `sys` package and exports `<sys/queue.h>` to ease
portability of external apps. `<os/queue.h>` was updated to use
`<sys/queue.h>`.

This fixes #972
